Add useFileRecord
